pyproject.toml
[tool.poetry]
name = "quote-django"
version = "0.1.0"
description = ""
authors = ["edwinbulter"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
django-cors-headers = "^4.6.0"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
